examples

+++ run_topopt.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# run with problem name as input var
+# for example run:
+# ./run_topopt.sh sphere_full.py
 
 ### SETUP ENVIRONMENT
 # by hand: env2lmod
@@ -9,14 +12,21 @@ module load gcc/4.8.5 cmake/3.16.5 openmpi/3.0.1 petsc/3.10.5 python/3.7.4
 # Print loaded modules
 module list
 
+# User input
+echo Press 1 for compiling...
+read var
+
 # compile
-rm -rf topoptlib.so
-rm -rf build
-mkdir build
-cd build
-cmake ..
-make
-cd ..
+if [ $var -eq 1 ]
+    then 
+        rm -rf topoptlib.so
+        rm -rf build
+        mkdir build
+        cd build
+        cmake ..
+        make
+        cd ..
+fi
 
 # RUN
 cd $SCRATCH/wrapped;
@@ -26,28 +36,21 @@ id=`date '+%Y%m%d_%H:%M:%S'`;
 
 mkdir $id
 cd $id
-#pwd
+pwd
 
-#path=../../../../home/thsmit/TopOpt_in_PETSc_wrapped_in_Python/problem_definition_file.py 
-#path=../../../../home/thsmit/TopOpt_in_PETSc_wrapped_in_Python/largescale.py
-#path=../../../../home/thsmit/TopOpt_in_PETSc_wrapped_in_Python/beam.py
-path=../../../../home/thsmit/TopOpt_in_PETSc_wrapped_in_Python/bracket.py
-#path=../../../../home/thsmit/TopOpt_in_PETSc_wrapped_in_Python/sphere.py
-#path=../../../../home/thsmit/TopOpt_in_PETSc_wrapped_in_Python/roof.py
+cp ../../../../home/thsmit/TopOpt_in_PETSc_wrapped_in_Python/topoptlib.so .
+cp ../../../../home/thsmit/TopOpt_in_PETSc_wrapped_in_Python/$1 .
 
 # ADJUSTABLE PARAMETERS
-#EULER_MEMORY="4000"
-#NCPU=32
-#WALL_TIME="8:00"
-
-# ADJUSTABLE PARAMETERS
-EULER_MEMORY="2000"
-NCPU=16
-WALL_TIME="4:00"
+#EULER_MEMORY="7000"
+#NCPU=16
+#WALL_TIME="48:00"
+EULER_MEMORY="1000"
+NCPU=4
+WALL_TIME="1:00"
 
 # FUNCTION CALL
-bsub -n ${NCPU} -W ${WALL_TIME} -R "rusage[mem=${EULER_MEMORY}]" mpirun -n ${NCPU} python $path 
-
+bsub -n ${NCPU} -W ${WALL_TIME} -R ib -R "rusage[mem=${EULER_MEMORY}]" mpirun -n ${NCPU} python $1
 
 # VIEW JOBS
 bjobs
